#26 move all logouts to the settings screen
