+++ app/build.gradle
@@ -170,7 +170,7 @@ dependencies {
     // Optional -- UI testing with UI Automator
     androidTestImplementation 'androidx.test.uiautomator:uiautomator:2.2.0'
 
-    androidTestImplementation 'com.android.support.test.espresso:espresso-idling-resource:3.0.2'
+    implementation 'com.android.support.test.espresso:espresso-idling-resource:3.0.2'
 
 
 }
