+++ app/src/bilbyblitz/java/fragments/setting/SettingFragment.java
@@ -1,10 +1,7 @@
 package fragments.setting;
 
-import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.StringRes;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,18 +11,13 @@ import android.widget.TextView;
 
 import org.greenrobot.eventbus.EventBus;
 
-import java.util.Formatter;
-import java.util.Locale;
-
 import activity.SingleFragmentActivity;
 import au.csiro.ozatlas.R;
-import au.csiro.ozatlas.manager.AtlasDialogManager;
+import au.csiro.ozatlas.fragments.settings.BaseSettingsFragment;
 import au.csiro.ozatlas.manager.Language;
 import au.csiro.ozatlas.manager.Utils;
 import au.csiro.ozatlas.model.Project;
-import base.BaseMainActivityFragment;
 import butterknife.BindView;
-import butterknife.ButterKnife;
 import butterknife.OnClick;
 import fragments.CustomSpinnerAdapter;
 import language.LanguageManager;
@@ -36,7 +28,7 @@ import static android.app.Activity.RESULT_OK;
  * Created by sad038 on 19/9/17.
  */
 
-public class SettingFragment extends BaseMainActivityFragment {
+public class SettingFragment extends BaseSettingsFragment {
 
     private final int PROJECT_LIST_REQUEST_CODE = 1;
     @BindView(R.id.languageSpinner)
@@ -47,10 +39,6 @@ public class SettingFragment extends BaseMainActivityFragment {
     TextView projectHeading;
     @BindView(R.id.languageHeading)
     TextView languageHeading;
-    @BindView(R.id.logout_settings_small_text)
-    TextView logoutSmallText;
-    @BindView(R.id.logout_settings_button)
-    View logoutButton;
 
     /*@BindView(R.id.offlineHeading)
     TextView offlineHeading;
@@ -97,17 +85,13 @@ public class SettingFragment extends BaseMainActivityFragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_settings, container, false);
-        setTitle(getString(R.string.setting));
-        ButterKnife.bind(this, view);
-
-        hideFloatingButton();
+        final View view = super.onCreateView(inflater, container, savedInstanceState);
 
         //language selection spinner setup
         String[] languages = getResources().getStringArray(R.array.bilby_language);
         //ArrayAdapter adapter = ArrayAdapter.createFromResource(getContext(), R.array.bilby_language, R.layout.item_textview);
         languageSpinner.setAdapter(new CustomSpinnerAdapter(getContext(), getResources().getStringArray(R.array.bilby_language), R.layout.item_textview));
-        //adapter.setDropDownViewResource(R.layout.dropdown_item_textview);
+
         int position = Utils.stringSearchInArray(languages, sharedPreferences.getLanguage());
         languageSpinner.setSelection(position == -1 ? 0 : position, false);
         languageSpinner.setOnItemSelectedListener(onLanguageSelectedListener);
@@ -115,39 +99,7 @@ public class SettingFragment extends BaseMainActivityFragment {
         //set the localized labels
         setLanguageValues(sharedPreferences.getLanguageEnumLanguage());
 
-        logoutSmallText.setText(getLoggedInShortMessage());
-
-        logoutButton.setOnClickListener(logoutView ->
-                AtlasDialogManager.alertBox(
-                        getActivity(),
-                        getString(R.string.logout_message),
-                        getString(R.string.logout_title),
-                        getString(R.string.logout_title),
-                        (dialog, which) -> launchLoginActivity()
-                )
-        );
-
         return view;
-    }
-
-    private String getLoggedInShortMessage() {
-        return formatResString(getContext(), R.string.logged_in_message, sharedPreferences.getUsername());
-    }
-
-    private static String formatResString(Context context, @StringRes int string, Object... args) {
-        final StringBuilder sb = new StringBuilder();
-        final Formatter f = new Formatter(sb, getCurrentLocale(context));
-        f.format(context.getString(string), args);
-        return sb.toString();
-    }
-
-    static Locale getCurrentLocale(Context context){
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N){
-            return context.getResources().getConfiguration().getLocales().get(0);
-        } else{
-            //noinspection deprecation
-            return context.getResources().getConfiguration().locale;
-        }
     }
 
     /**
