+++ app/src/bilbyblitz/res/layout/header_list.xml
@@ -37,24 +37,6 @@
             android:textSize="@dimen/bigger_input_font_size"
             android:textStyle="bold" />
 
-<!--        <TextView-->
-<!--            android:id="@+id/logoutButton"-->
-<!--            style="@style/TextAppearance.AppCompat.Small"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginTop="12dp"-->
-<!--            android:background="@android:color/white"-->
-<!--            android:foreground="?attr/selectableItemBackground"-->
-<!--            android:gravity="center"-->
-<!--            android:paddingBottom="6dp"-->
-<!--            android:paddingLeft="16dp"-->
-<!--            android:paddingRight="16dp"-->
-<!--            android:paddingTop="6dp"-->
-<!--            android:text="@string/logout_title"-->
-<!--            android:textAllCaps="true"-->
-<!--            android:textColor="@color/colorPrimary"-->
-<!--            android:textSize="@dimen/smaller_input_font_size"-->
-<!--            android:textStyle="bold" />-->
     </LinearLayout>
 
 </RelativeLayout>
