+++ app/src/bilbyblitz/res/menu/activity_main_drawer.xml
@@ -36,16 +36,6 @@
 
     </group>
 
-
-<!--    <item android:title="Account">-->
-<!--        <menu>-->
-<!--            <item-->
-<!--                android:id="@+id/nav_logout"-->
-<!--                android:icon="@drawable/ic_exit_to_app_black_24dp"-->
-<!--                android:title="@string/logout" />-->
-<!--        </menu>-->
-<!--    </item>-->
-
     <group
         android:id="@+id/contact"
         android:checkableBehavior="none">
