+++ app/src/bilbyblitz/res/values/strings.xml
@@ -38,7 +38,6 @@
     <string name="partners">Partners</string>
     <string name="help">Help</string>
     <string name="update_location">Update Location</string>
-    <string name="setting">Settings</string>
     <string name="offline_setting">Offline Settings</string>
     <string name="no_animal_record">No sightings have been recorded yet</string>
     <string name="edit_track">Edit Track</string>
