+++ app/src/biocollect/java/fragments/HomePageFragment.java
@@ -1,6 +1,5 @@
 package fragments;
 
-import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -15,7 +14,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import au.csiro.ozatlas.R;
-import au.csiro.ozatlas.manager.AtlasDialogManager;
 import au.csiro.ozatlas.model.HomePageListItem;
 import base.BaseMainActivityFragment;
 import butterknife.BindView;
@@ -75,23 +73,11 @@ public class HomePageFragment extends BaseMainActivityFragment {
      */
     private void setupHeader(View header) {
         TextView nameTV = (TextView) header.findViewById(R.id.name);
-//        TextView logoutTV = (TextView) header.findViewById(R.id.logoutButton);
         String name = sharedPreferences.getUserDisplayName();
         if (name == null || name.equals(""))
             nameTV.setText(getString(R.string.welcome_message, getString(R.string.app_name)));
         else
             nameTV.setText(getString(R.string.good_day_message, sharedPreferences.getUserDisplayName()));
-//        logoutTV.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                AtlasDialogManager.alertBox(getActivity(), getString(R.string.logout_message), getString(R.string.logout_title), getString(R.string.logout_title), new DialogInterface.OnClickListener() {
-//                    @Override
-//                    public void onClick(DialogInterface dialog, int which) {
-//                        launchLoginActivity();
-//                    }
-//                });
-//            }
-//        });
     }
 
     /**
