+++ app/src/biocollect/res/layout/header_list.xml
@@ -37,22 +37,6 @@
             android:textColor="@color/header_text_color"
             android:textStyle="bold" />
 
-<!--        <TextView-->
-<!--            android:id="@+id/logoutButton"-->
-<!--            style="@style/TextAppearance.AppCompat.Small"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginTop="12dp"-->
-<!--            android:background="@drawable/button_background"-->
-<!--            android:gravity="center"-->
-<!--            android:paddingBottom="6dp"-->
-<!--            android:paddingLeft="16dp"-->
-<!--            android:paddingRight="16dp"-->
-<!--            android:paddingTop="6dp"-->
-<!--            android:text="@string/logout_title"-->
-<!--            android:textAllCaps="true"-->
-<!--            android:textColor="@android:color/white"-->
-<!--            android:textStyle="bold" />-->
     </LinearLayout>
 
 </RelativeLayout>
