--- app/src/main/java/au/csiro/ozatlas/fragments/SettingsFragment.java
@@ -1,10 +1,0 @@
-package au.csiro.ozatlas.fragments;
-
-import android.preference.PreferenceFragment;
-
-// TODO Upgrade to Android jetpack
-public class SettingsFragment extends PreferenceFragment {
-
-
-
-}
