+++ app/src/main/res/values/style.xml
@@ -7,5 +7,19 @@
         <item name="android:textSize">@dimen/default_input_font_size</item>
     </style>
 
+    <!--input label style for settings list item-->
+    <style name="SettingsLabel">
+        <item name="android:singleLine">true</item>
+        <item name="android:textAppearance">?android:attr/textAppearanceLarge</item>
+        <item name="android:ellipsize">marquee</item>
+        <item name="android:fadingEdge">horizontal</item>
+    </style>
+
+    <style name="SettingsSubLabel">
+        <item name="android:textAppearance">?android:textAppearanceSmall</item>
+        <item name="android:textColor">?android:attr/textColorSecondary</item>
+        <item name="android:maxLines">4</item>
+    </style>
+
     <style name="Toolbar.TitleText" parent="TextAppearance.Widget.AppCompat.Toolbar.Title"></style>
 </resources>
