--- app/src/main/res/xml/preferences.xml
@@ -1,3 +1,0 @@
-<PreferenceScreen>
-
-</PreferenceScreen>
