+++ app/src/ozatlas/java/activity/MainActivity.java
@@ -4,7 +4,6 @@ import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.graphics.Color;
@@ -30,7 +29,7 @@ import au.csiro.ozatlas.BuildConfig;
 import au.csiro.ozatlas.R;
 import au.csiro.ozatlas.base.BaseActivity;
 import au.csiro.ozatlas.base.MainActivityFragmentListener;
-import au.csiro.ozatlas.manager.AtlasDialogManager;
+import au.csiro.ozatlas.fragments.settings.BaseSettingsFragment;
 import au.csiro.ozatlas.manager.AtlasManager;
 import fragments.AddSightingFragment;
 import fragments.DraftSightingListFragment;
@@ -135,15 +134,9 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
         // Handle navigation view item clicks here.
         int id = item.getItemId();
 
-//        if (id == R.id.nav_logout) {
-//            AtlasDialogManager.alertBox(this, getString(R.string.logout_message), getString(R.string.logout_title), getString(R.string.logout_title), new DialogInterface.OnClickListener() {
-//                @Override
-//                public void onClick(DialogInterface dialog, int which) {
-//                    launchLoginActivity();
-//                }
-//            });
-//        } else
-        if (id == R.id.nav_add) {
+        if (id == R.id.nav_setting) {
+            getSupportFragmentManager().beginTransaction().replace(R.id.fragmentHolder, new BaseSettingsFragment()).commit();
+        } else if (id == R.id.nav_add) {
             getSupportFragmentManager().beginTransaction().replace(R.id.fragmentHolder, new AddSightingFragment()).commit();
         } else if (id == R.id.nav_all_sighting) {
             getSupportFragmentManager().beginTransaction().replace(R.id.fragmentHolder, new SightingListFragment()).commit();
