+++ app/src/ozatlas/java/fragments/HomePageFragment.java
@@ -86,14 +86,7 @@ public class HomePageFragment extends BaseMainActivityFragment {
      */
     private void setupHeader(View header) {
         TextView name = (TextView) header.findViewById(R.id.name);
-//        TextView logout = (TextView) header.findViewById(R.id.logoutButton);
         name.setText(getString(R.string.good_day_message, sharedPreferences.getUserDisplayName()));
-//        logout.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                launchLoginActivity();
-//            }
-//        });
     }
 
     /**
