+++ build.gradle
@@ -11,8 +11,7 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:3.5.0'
         classpath 'com.google.gms:google-services:4.3.2'
-//        classpath 'io.fabric.tools:gradle:1.25.4'
-        classpath 'io.fabric.tools:gradle:1.+'
+        classpath 'io.fabric.tools:gradle:1.31.1'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
